Add pgbouncer support
Pgbouncer support has been added in a way the mirrors how pgpool is
implemented. There are still rough edges and these include how to name
the pool themselves. The current names are <cluster name> and <cluster
name>-replica. The other issue is for admin purposes the primaryuser
for the cluster is used, and instead there should probably be a user
only on pgbouncer. However all these issues can be handled by using
predefined secrets.