Moved when to do create .pgouser file
I move the directions to create the .pgouser file up to the operator security section since this is where we told users where the default set of roles and users are defined.  The user will get an error when trying to execute pgo version in the pgo cli section because the creation of the .pgouser file was not done yet and was told how to create the file later in the document.