Add environment vars loading for username/password (#713)
These feature allows for users to store credentials in a manner other
than a plaintext file. This method is preferred in some environments
where writing credentials to a plaintext file is not permitted.

PGOUSERNAME and PGOUSERPASS were chosen to be in line with PGOUSER but
be more specific. The idea that if either are set then the PGOUSER file
will be ignored is deliberate. These vars are part of a new feature
only and as such using either is interpreted as an attempt to use both.
The decision to rename pgouserenvvar to pgoUserFileEnvVar was made to
reduce confusion between the username/password auth and file based auth.
The function added to apiservermsgs/common.go was added in place of
changing SessionCredentials to *SessionCredentials and performing a nil
check.