Fix pgbadger target
Allow pgbadger instances were attempting to target the primary's pgdata
dir. This patch adds a single argument to GetBadgerAddon that allows
for the target to be passed in. The restore was also adjusted with this.
I have not seen the issue on restore, but based off the code it looks
like restored clusters will have the same issue. Also I believe that
collect has this same issue.