Update to how CentOS containers utilize the upstream PostgreSQL repositories.
The upstream PostgreSQL repositories (PGDG) updated how RPMs and their dependencies
are ingested, which caused some of our previous configured way of download the RPMS to
fail.

This commit adjusts the way the RPMs are downloaded. Instead of disabling the repositories
we do not need, we disable _all_ the repositories and only enable the one we need. The yum
calls were updated in the pgo-backrest dockerfile to have better readability and stability.
This dockerfile will now fail if any packages do not exist. Previously this was
accomplished by running separate yum commands for each but is now done with in one command
by using the setopt flag.

For example, for the PostgreSQL 11 RPMs, we would disable _all_ of the repositories
(`pgdg*`) and only enable the one with the PostgreSQL 11 RPMs (`pgdg11`)