Remove references to node_exporter in config, installation, & docs.
The node_exporter container is no longer a part of the Crunchy
Container suite per:

https://github.com/CrunchyData/crunchy-containers/commit/87107f4d320a1705e30c44719d848ffb951dca64

This is due to Kubernetes and Kube-derived builds providing their own
node exporters in each Kubelet that provides their own metrics via
cAdvisor. Additionally, node_exporter would provide metrics on an
entire node vs. what was in the container itself, which might not
provide the requisite info for monitoring and diagnosing issues in
a particular container.