Enables the bypassing of authentication for health check endpoint
This allows the health check endpoint to be accessible to large-scale monitoring tools.

Prior to this, all URL routes require mutual TLS authentication and all but the /health route
requires HTTP Basic authentication. The /health route potentially prompts for HTTP Basic, but
does not enforce it.

When the `NOAUTH_ROUTES` environment variable is correctly configured, the server's TLS
configuration is reduced to verifying presented client certificates if presented vice both
requiring and verifying client certificates.

For all protected routes, the existence of verified client certificates is confirmed or a
"HTTP Forbidden" status is returned.

For noauth routes, that verified certificate check is skipped.

At this time, only the /health endpoint is allowed to be a noauth route. Enabling future endpoints
will require they remove dependencies on the username returned from HTTP Basic
authentication. /health already had no dependency and no validation prior to this change.