Improvements to Postgres Operator Ansible Inventory Validation
These changes improve the inventory validation to better ensure
the appropriate values are set before completing an Openshift or
Kubernetes Operator install. Most of the checking now happens during
preflight, meaning the problems are identified earlier.
Additionally, there are checks to see that the required values are not
just set but are also not null.