Updates operator to set service ports as env vars.
Currently pgbadger and postgres-exporter service ports are used
to create the service but not passed into the container. This
update changes the deployment json files so that the service ports
are set in the environment of the container. The ports defined in
the pgo-config are passed into the deployment templates and used
to create the containers.