removes unused Clusteridentifier field
The Clusteridentifier field was added for a future event client
requirement to uniquely identify the underlying cluster.
The current implementation does not return Clusteridentifier
content with the message as nothing is populating the pg-cluster-id
is not being set.

Any direct usage of the field in dependent code will fail compilation.