Update Postgres Operator Upgrade Procedures to Match Updated Code Behavior
In line with commit "Update cluster upgrade"
https://github.com/CrunchyData/postgres-operator/pull/1027
The upgrade procedures have been updated.

For the PGO 4.0.1 to 4.1.0 upgrade procedure, the new
functionality is referenced in place of the previous
manner of executing the upgrade.

For the PGO 3.5 Minor Version upgrade procedure, the
upgrade procedure has been updated so that the behavior
will match the updated 4.1.0 code change.