enforces required env variables and adds docs (#1047)
add-targeted-namespace.sh has a sinister failure case if
PGO_INSTALLATION_NAME is unset in that the pgo-installation-name label
does not get set correctly. This causes later failures in selectors
built upon the assumption of that label's existence.

This change updates add-targeted-namespace.sh to fail if its expected
environment variables do not exist and updates the references to the
script to warn about needing to ensure its definition.