Postgres Operator Upgrade updates (#1051)
* Postgres Operator Upgrade updates

An updated procedure for upgrading the Postgres Operator from
version 4.0.1 to 4.1 is required to account for recent configuration
updates and environmental changes. These changes left certain
Operator features nonfunctional.  The new upgrade procedure
is designed to ensure all applicable changes are accounted for in
the most convenient way possible.

* Updated the PGO 4.0.1 to 4.1.0 procedure to include a missing step
outlining the process to recreate the 4.0.1 clusters in the new
4.1.0 environment.

Also added a reference to the pgo scaledown command's help for use
when needing to scale down the existing PGO 4.0.1 pgclusters.

* Added a separate upgrade procedure for users using the Ansible
installation method for their Operator installation. Procedure is
designed to be inclusive. Other updates reflect renaming of the
existing procedure to specify it is for Bash installations, and
an update to relevant links.