updated broken links (#1053)
update links in the documenation that were giving 404 errors. I updated the design link in the configuration.md file. I also updated the pgo cli sntax and examples, design overview, and developing links in the index.md file. I updated the github repository link for contributing and fixed the link for namespace formatting reference on namespace.md and olm.md