Rewrite the Makefile for more consistency
The makefile was previously hard to read and follow. In order to make
future changes easier, a few changes were made for consistency. They are
as follows:

- All targets are marked as .PHONY directly at the top of the file. The
reason for this is all the targets are phony so the intent is to make
that obvious to the user upfront.

- Added only one black line between each target, but there are 2 black
lines before sections denoted by a `#======= * =======`.

- Reordered targets so that similar ones are next to each other.

- Made targets follow the standard of a single space after each colon.
Previously there were sometimes tabs and sometimes spaces.

- Removed calls to make from inside the Makefile. Targets will now just
call intended targets instead of executing a new make command.

- Added comments to most targets that would be used by outside users to
indicate what they will do.

A few features were added to help with consistency and ease of use:

- There is now a help target that will display lines that have a
target defined followed by a `## <message>`. This allows for `$ make
help` to be run and general idea of what commands do will be.

- The new help command has been set as the default command for make if
no target is specified.

- Added an equivalent clean-*-image for each image target.

- Added a global clean-images that calls all the previously mentioned
clean-*-images (NOTE: buildah is not parallel safe)