This commit integrates the 'crunchy-postgres-ha' container into the PGO.  Now when clusters are created, as well as when additional replicas are added to existing clusters, the 'crunchy-postgres-ha' container is utilized instead of the 'crunchy-postgres' container.  This effectively introduces integrates Patroni into the PGO architecture, which will be leveraged to bootstrap, manage and ensure the availability of any PG databases and clusters created by the PGO.
With the introduction of the 'crunchy-postgres-ha' container,
an additional service account named 'pgo-pg' is now created
and utilized by any PG deployments, with any permissions
required by Patroni to create and modify the various artifacts
in Kubernetes required to support the Patroni solution.  Also,
Patroni environment variables are utilized to configure the
'crunchy-postgres-ha' container, along with a Patroni
configuration file and applicable Kubernetes secrets.

Being that this change represents a major change to the
PGO architecture, this is considered a breaking change
that will not be backwards compatible with an PG clusters
previously created by the PGO.  Please note that this
commit also does not yet deprecate or update any failover
logic, as required to fully leverage the HA capabilities
provided by Patroni in the  'crunchy-postgres-ha' container
(these updates will be provided in a subsequent commit).