Add disable TLS functionality to the PGO Client
The Postgres Operator's apiserver currently supports the
disabling of TLS authentication when receiving client connections.
However, the PGO client did not have a method for disabling
the necessary resource checks that are then unnecessary, meaning
the unneeded certificates and keys were still required to be
present.

This update allows the PGO client to skip the relevant configuration
setup when using an unauthenticated connection. This can be done in
one of two ways. Either by an environment variable:

DISABLE_TLS=true

(As is used by the Postgres Operator to disable TLS when performing
a Bash install)

Or by using the '--disable-tls' flag with the PGO client, for example:

pgo version --disable-tls

In the event that the DISABLE_TLS variable is set to 'false' and the
--disable-tls is used, the flag is authoritative.

Additionally, this commit includes documenationupdates to the main
TLS configuration page to explain how to properly set the new PGO
CLI configuration variables depending on installation method.

Notes with links to the main page were also added to both
the Bash and Ansible installation sections.

Issue: [ch6126]