The add-targeted-namespace.sh script now creates the Service Account, Role and Role Binding required by the 'crunchy-postgres-ha' container.
The PGHA_PRIMARY_HOST environment variable is
now properly set on any replica deployments, which
ensures the Primary pod is created and initialized prior
to the replicas.  This enables the proper orchestration
and placement of Primary and Replica pods.

The PG cluster deployments and associated services
now include the Patroni port (8009).