Modify PostgreSQL cluster test to be more efficient.
Previously, the PostgreSQL cluster test (from the command line, known as
`pgo test`) would check if each individual user available to a PostgreSQL
cluster could authenticate with its credentials. While this certainly
determines that one can make a connection to a PostgreSQL database managed by
the Operator, it also presents several issues:

- If one has many users available to the Operator, the number of connections
made can rapidly multiply. Imagine 10 users across 5 replicas
- This also means that the user's credentials must be stored in the Operator.
Currently this is a plaintext password. If the connections are _not_ being made
over SSL, this creates the risk of password leakage.

The test now performs two types of checks: instance (or pod) readiness checks,
and endpoint availability. Both of these are designed to check if a PostgreSQL
instance an be connected to and reached from the outside world. Specifically:

- The pod readiness check calls `pg_isready` which can quickly determine if a
PostgreSQL is accepting connections
- Check if the service endpoints responsible for serving the PostgreSQL are
reachable.

If both of the checks pass, then a PostgreSQL cluster is reachable.

This commit also does the following:

- BREAKING: Modify the API endpoint that is consumed by `pgo test` to deliver
the content in a structure manner
- BREAKING: Remove the "TestCluster" event. This is not a lifecycle event.
- Introduce the "Endpoints" Kubernetes construct to the `kubeapi` package
- Some refactoring, cleanup, and improved documentation around the connectivity
tests

Issue: [ch5712]