Modify behavior of cluster deletion to remove all data by default.
Previously, calling `pgo delete cluster` causes a cluster to be deleted, but
keeps the backups and data PVCs available.

However, this can lead to some unexpected behavior such as users in a
"recreated" cluster are unable to access their accounts as their credentials
have changed.

Generally, the user expectation when deleting a PostgreSQL cluster is that all
of their data is deleted. If a user wants to retain data from a cluster,
the standard method is to take a full backup (physical or logical) of all the
data in that cluster. When needed, the user can restore that data to a new
PostgreSQL cluster.

Now, when a PostgreSQL cluster is deleted, **all data and backups are deleted
by default**. Two new flags are introduced to the `pgo delete cluster` command
to manage this new behavior:

- `--keep-backups`, which allows backups to be retained and user at a later
time.
- `--keep-data`, which allows the data to be retained. This command is
immediately deprecated.

The `-b`, `--delete-backups`, `-d`, `--delete-data` flags are deprecated,
though they effectively amount to a no-op.

Some extraneous formatting may have also made its way in, as a result of
my automatically running `gofmt` on any file I modified.

Issue: [ch6206]