Adds comments to API code to generate swagger docs
This commit adds godoc style comments that are used along with
go-swagger to generate a swagger openAPI spec. The spec can be
used render a webpage or pdf using other swagger tools.

No current functionality is changed and only comments are added
in the packages related to the api endpoints and server messages.