Set the pgo-backrest-repo container to run as OS user "2000"
A container instantiated with the "runAsNonRoot" is unable to
verify that the "pgbackrest"  user is a non-root OS user. The fix for
this is to explicitly specify the numeric ID associated with the user
account in order for "runAsNonRoot" to verify that this is not a root
account.

Issue: [ch6248]