Ensure Kubernetes Secrets associated with backups are deleted when specified.
During a PostgreSQL delete operation where the user explicitly specifies to
delete the backups `e.g. pgo delete cluster clustername --delete-backups`, the
`<clustername>-pgbackrest-repo-config` secret was not being deleted. Given all
the backups for that cluster are destroyed, the secret is effectively moot and
should be removed.

This ensures that the `<clustername>-pgbackrest-repo-config` secret is removed
when `--delete-backups` is specified. However, if only `--delete-data` is
specified but not `--delete-backups`, the aforementioned secret remains.

Issue: [ch6254]