Updates the cluster initialization process as needed to support pgBackRest replica creation.  This means that in addition to creating the pgBackRest stanza following initialization of the primary, a full backup is also taken (the backup is executed following a successful stanza creation).  This full backup can then be utilized to initialize any replicas in the cluster using the pgBackRest initialization process included in the containers.
In order to support this initialization process, any
pgreplica CR's created prior to the successful
initialization of the primary are placed in a pending
state until the cluster is fully initialized (i.e. the stanza
and full backup have been created).  Then, once the
cluster is initialized, and update to the pgreplica's is
triggered via an annotation, triggering the creation
of the replicas.  A 'pgcluster Initialized' state is now
included for the pgcluster CR, and is therefore utilized
to determine if the cluster is fully initialized (e.g. in
order to trigger replica creation).

Also included are updates to the pgBackRest restore
workflow as needed to support the new
crunchy-postgres-ha and crunchy-postgres-gis containers.
Now when a restore is initiated, all replicas are deleted
in addition to the primary.  Then, once the restored
primary DB comes back online, the replicas are recreated
using any existing pgreplica CR's.  This essentially establishes
a fully working cluster following a successful restore,
which includes all desired replicas and autofailover
enabled by default.

Additionally, it should be noted that any  PVC's previously
utilized by those replicas are reused for those same replicas
in the restored cluster.  Also, the pgcluster CR now shows
a 'pgcluster Restoring' state during a restore, and any
pgreplica will show a "pgreplica Pending restore" state
while the restore is underway.