Replaces SSH with kubectl for communicating between the pgBackRest dedicated repository host and the primary PG database.  Specifically, a script has been created that translates any SSH commands issued by pgBackRest into the proper kubectl command.  pgBackRest is then configured to call this script when SSH is required, instead of calling SSH directly.  This allows kubectl to be utilized as a drop-in replacement for SSH, and therefore prevents the need to install and run SSH on any containers created by the PGO (for instance, SSH no longer needs to be enabled on the dedicated repository host nor the PG primary pod).
However, it should be noted that SSH can still be enabled
in any backrest container if necessary while the kubectl
solution is fully verified and validated.  Therefore, at this
time SSH is still installed installed in any backrest containers.
Additionally, kubectl and the associated yum repo are now
installed in any backrest containers to support the use of
kubectl instead of SSH.