Update documentation for Ansible installer requirements.
The required configuration parameters for the Ansible installer were beneath the
table that listed out most of the available parameters, as well as helpful
information about storage configuration. This could make it challenging to set
up the inventory file for the install.

This reorganizes the information so that the required configuration variables
and other information necessary to configure the installer are moved up higher
in the documentation, as well as including a `Required` column in the table
that details what all the parameters do.

Additionally, not all of the required parameters were documented, so this adds
the missing ones to the documentation.

Issue: [ch6100]