Add `--keep-data` flag to `pgo scaledown`
Previously, `pgo scaledown` would by default keep the PVC claim for a
specific cluster, which runs counter to expected behavior when a node is
scaled down, and would often create many dangling PVCs.

This changes the default behavior of `pgo scaledown` to delete the PVC
for the PostgreSQL instance that is being removed from the cluster.

This also  introduces the `--keep-data` flag to `pgo scaledown` which
allows a user to indicate if they wish to keep the PVC with the data
from the replica.

The `--delete-data` flag for `pgo scaledown` is now deprecrated.

Issue: [ch6378]