Enable pgBouncer to connect to additional database
The previous configuration for pgBouncer would not allow it to work with
any PostgreSQL databases that were created after the initial deployment
of the Operator.

This leverages the "catch-all" feature of pgBouncer (i.e. "*") to point
to any database that is associated with the primary PostgreSQL service.

Additionally, this installs the pgBouncer "get_auth" function into
the "template1" database to ensure that it is available in any newly
created databases.

For existing deployments of the Operator, one needs to update their
"pgbouncer.ini" files. The installation templates for the pgBouncer
files are updated to support this chaange.

This change continues to work with the changes to the HA system as it
references the primary service only.

Issue: [ch6204]
Issue: [ch5472]