This commit effectively removes the custom PGO autofail solution from the PGO.  In its place, Patroni (contained within the crunchy-postgres-ha and crunchy-postgres-gis-ha containers) will now be responsible for ensuring any PG clusters created by the PGO are highly available.  This includes automatically failing over to (i.e. promoting) a replica in the event of a failure in the primary PG database.  With this change, once bootstrapped PG clusters are now self-sufficient and able to self-heal without relying directly on the PGO itself and its availability.
In support of the transition to this new high availability
solution are the following notable design changes:

- Autofail is now enabled by default for any newly created
cluster, and a warning is returned from the API server in the
event that the new cluster contains zero replicas (specifically
indicating that at least one replica is needed to leverage
autofail functionality
- Disabling/enabling autofail means "pausing"/"resuming"
Patroni (i.e. attaching/detaching Patroni from the cluster).
This is accompished by setting/removing the
{"pause": true/false/null} key in the cluster's "config"
configMap.
- With autofail now enabled by default, the "autofail" flag for
the "pgo create" and "pgo update" commands has been
changed to "disable-autofail".  When set, this option disables
autofail in the newly created cluster (however, it can still
later be enabled).  Similarly, the 'Autofail' option
in pgo.yaml has been changed to 'DisableAutofail'
- The AutofailReplaceReplica failover option has been
removed from the API server and CLI (i.e. the
"autofail-replace-replica" flag, since it is no longer applicable
to the current design.  Specifically, with the
'crunchy-postgres-ha'  container, the old primary will now
automatically be turned into a replica that will follow
the new primary (with pg_rewind utilized if needed)

Manual failover capabilities are now also implemented in
support of he 'crunchy-postgres-ha' container.  Specifically,
the proper Patroni REST endpoint is now called via the exec
endpoint in order to failover to a replica.  Additionally,
when any failover occurs, whether automatic or manual, the
role label change on the new primary from "replica" to "master"
is detected, and the pgBackRest dedicated repository host is
updated to point to the proper "PGDATA" directory for the promoted
DB node.  And finally, in support of proper replica selection
for failover functionality, label selectors have been updated for
the "--query" options for "pgo scaledown" and "pgo failover" to
return the proper cluster replicas (including after a failover),
while "pgo show cluster" also now properly displays primary and
replicas following both before and after failover.

Also includes documentation changes required to support the full
replacement of the custom PGO HA solution with the HA solution provided
by the 'crunchy-postgres-ha' container and Patroni.