Revert "Adds the kube-ssh-wrapper.sh script for using"
This reverts commit 3bd9abbd87729d249739ca165ee3fa830e5a0140.

This effectively removes the kube-ssh-wrapper.sh wrapper script
utilized as a drop-in replacement for SSH for use by pgBackRest,
as well as any configuration settings in place for use by this
script.  Additionally, this reverts various pgBackRest
configuration settings in place to disable SSH and utilize the
kube-ssh-wrapper.sh script to once again enable and utilize SSH
for all pgBackRest interactions between primary DB pods and
the dedicated repository host.