refactors apiserver route registration
Moves route registration to a utility function provided by
the package itself, reducing direct dependencies on service
packages and general noise in the main package.

NB - to resolve circular dependencies, this route population
was pushed to a lower-level package

This change further paves the way for greater testibility
by allowing binding the api routes to a test router.