removes extraneous file read
This appears to be left over from a time where the files were
loaded sideband. However, this file was *just* written by
apiserver.WriteTLS and a "read to check" is unnecessary here.

Furthermore, ListenAndServeTLS requires the key from this
file to start correctly, so an unreadable file state will
not remain hidden.