moves TLS server preparation into TLS-only block
The code to prepare the TLS configuration is meaningless in
the context of DISABLE_TLS being set. This change avoids
unnecessary work for TLS-disabled configurations and prevents
startup failures in TLS-disabled mode from TLS preparation.

To improve logging meaning, logging listening status is moved
to just before the call to Listen*