Remove PGcluster Configmaps on Delete
This commit adds the function removeClusterConfigmaps which deletes the three
configmaps that are now created for each cluster. They consist of two that are
created by Patroni when it initializes a new cluster:
 <cluster-name>-leader (stores data pertinent to the leader election process)
 <cluster-name>-config (stores global/cluster-wide configuration settings)
and one that contains a default Patroni configuration file:
 <cluster-name>-pgha-default-config (stores a Patroni config file in YAML format)
These configmaps are specific to the cluster and, therefore, do not need to
persist after cluster deletion.