Ensure `pgo test` shows instances in "unknown" state
When a PostgreSQL instance is in an indeterminate state, e.g. during
initialization or when healing after a failover event, the type of
instance is unknown. However, `pgo test`, and by extension the test
endpoint in the `apiserver`, were not returning the status of the
instances that were in this state.

This patch corrects this by allowing instances to have a type of
"unknown".

Issue: [ch5479]