prevents misleading error-level logs
The apiserver and nsqd processes do not gracefully handle the
empty TCP connections generated by kubelet as part of probe
execution, causing misleading error messages to accumulate
in the logs at the probe polling intervals.

For the nsqd process, there exists an HTTP endpoint that can
be used for liveness testing, but it not useful for readiness
testing as the `/ping` endpoint does not reflect readiness
for connections on port 4150. Thus, the liveness test is moved
to an HTTP test of the `/ping` endpoint and the readiness test
is removed.

Furthermore, each access to `/ping` is logged as an INFO level
log in the nsqd log files. While this is not as misleading as
ERROR level logging, it still represents unnecessary logging
and the log level for the nsqd process has been moved up to
WARNING.

For the pgo-apiserver, a new lightweight route is added to
the conventional location of `/healthz` with a simple `ok`
response with the HTTP 200.

As a result, NOAUTH_ROUTES no longer optionally activates
deferred certificate checking - it is required to support
`/healthz` as kubelet does not present a certificate when
performing HTTPS checks.

NOAUTH_ROUTES may still be used to enable unauthenticated
access to the existing whitelist of routes.

[ch6359]
[ch6452]