The PGHA_PRIMARY_HOST environment variable has been removed from the cluster-deployment.json template, which means it will no longer be included in any replicas created by the PGO using this template. This environment variable, while useful for orchestrating primary and replica creation in a containers environment such as Docker, is not needed when utilizing the crunchy-postgres-ha container with the PGO. This is because the PGO is able to leverage Kubernetes to orchestrate primary and replica creation.
The removal of this variable will prevent a former replica
from waiting indefinitely for the original primary to start
in the event that it is rebooted following a failure and the
original primary is unable to come online.  This could occur
during an unscheduled failover, or when one is expected, e.g.
when the primary is upgraded during a minor upgrade).
Therefore, removing this setting ensures that any replica
can successfully become primary following a failure.