This commit updates the mechanism for patching (i.e. updating) the image name for the database, as well as any sidecar containers (crunchy-collect, crunchy-pgbadger and/or crunchyadm) included in a primary or replica Deployment, during the minor upgrade of a PG cluster.  Now instead of using the 'json' patch type to perform a JSON patch in accordance with RFC 6902 to upgrade the image name for any/all containers included in a PG primary or replica Deployment, a Kubernetes strategic merge patch is utilized instead. This change provides a cleaner mechanism for updating container image names, namely because the number of containers included in a PG Deployment could vary based on cluster configuration, i.e the same sidecar containers might not always be enabled in each cluster.
Specifically, being that the 'json' patch type requires
the image name to be updated via a positional array in
the Deployment spec, it can be challenging to identify
exactly which sidecars are included at each position
within the "containers" array in order to ensure the
proper image name is specified when performing the patch
request (i.e. the position of each sidecar could change
across clusters depending on what sidecars are enabled,
or if additional sidecars are either added or removed in
the future).  However, by utilizing a strategic merge
patch, it is possible to update each container's image
name using the name of the container only, making the
specific position of the container in the "containers"
array irrelevant, and therefore facilitating the addition
and/or removal of sidecar containers in the future.

If autofail is disabled when performing a minor upgrade,
the primary and all replicas are patched at the same time,
and the upgrade task is immediately marked as complete.
This is because when performing a minor upgrade with autofail
disabled, once the upgrade is complete and new pods have been
created for the primary and each replica, any "database"
containers will remain in an unready state, i.e. readiness
probes will fail for those containers (and will continue to
do so indefinitely).  This is because when autofail is
disabled (i.e. Patroni is "paused"), Patroni will not attempt
to start the PG database on a primary or replica.  Therefore,
the user must re-enable autofail following a minor upgrade
during which autofail was disabled in order to fully bring
the cluster (which includes the primary and all replicas) back
online.