Function UpgradeAutofailDisabled has been renamed to UpgradeWithAutofailDisabled.
Additionally, a similar UpgradeWithAutofailEnabled
function has been created, which is called when a new
upgrade has been initiated for the cluster, and autofail
is enabled for that cluster.  Since an upgrade initiated
with autofail enabled triggers a sequential upgrade of
all replicas and then the primary, this function simply
calls the ProcessNextUpgradeItem function, which is
responsible for processing the next (and in this case
the first) Deployment in the sequence of Deployments
(primary and/or replicas) requiring an update.

This clearly defines and differentiates the possible minor
upgrade paths,as determined by whether or not autofail is
enabled in the cluster.