Fix failure for when a replica is not utilizing streaming replication
PostgreSQL replicas can replay transactions directly from a WAL archive, such as
the one provided by the pgBackRest in the Operator. As such, it may not be able
to connect to a primary.

However, the `pgo scaledown --query` command was unable to process this case.
This patch ensures that it can display the current LSN positions even if the
replica is replaying from the WAL log and is not utilizing streaming
replication.