Update `pgo failover --query` output
While `pgo failover --query` provided information that could help a
user make a decision about which instance in a PostgreSQL cluster they
would want to fail over to, it could be challenging to decipher as not
everyone is versed in PostgreSQL log sequence numbers (LSN).

This changes leverages the information that is being collected by the
new HA system in order to provide better feedback about the status of
the replicas available. This information includes:

- NAME: the name of the replica
- STATUS: the current status of the replica
- NODE: the node that the replica resides on
- REPLICATION LAG: how far behind the replica is behind the primary

Additionally, "TIMELINE" is available in the API but is not currently
displayed to the user.

This also includes a refactor of the previous change that added more
information to the `pgo scale --query` so that the changes could be
consolidated. From a user standpoint, both of these commands require
similar information to make informed decisions, so it makes sense to
consolidate the heavy lifting.

This also cleans out some now unused code.

Issue: [ch6494]