Get the current list of PVCs to delete before deleting them
Commit 34341bb introduced a feature to clear out the ConfigMaps that are created
with an associated cluster. However, presently there is a failure case where if
a PVC fails to delete, that this line refuses to get executed.

The root of the issue is that PVCs can be deleted at different times during the
"rmdata" job, we need to refresh the list of PVCs to be deleted before doing the
final pass through them.

Regardless, it's not a bad idea to clear out the ConfigMaps before removing the
final PVCs, so we can move this line of code up.

Issue: [ch6324]