Get the current list of PVCs to delete before deleting them
There is a failure case where if a PVC fails to delete, that this line refuses
to get executed.

The root of the issue is that PVCs can be deleted at different times during the
"rmdata" job, we need to refresh the list of PVCs to be deleted before doing the
final pass through them.

Issue: [ch6499]