Add ability to clone a PostgreSQL cluster
Perhaps to be commonly know as `pgo clone`, this allows a PostgreSQL
Operator user to safely clone a PostgreSQL cluster. One is required to
use pgBackRest on their source cluster, but this is a requirement
starting with PostgreSQL Operator v4.2.0.

Presently, we only support doing a full clone, i.e. the current state
of the cluster we are cloning is mapped over. The intention is to also
allow options such as a "point-in-time" clone, where the cluster is
cloned and restored to a particular point in time.

The clone works as such:

1. One initiates a clone using `pgo clone oldcluster newcluster`, where
"oldcluster" is the PostgreSQL cluster with the source data and
"newcluster" is the new PostgreSQL cluster we want to create

2. A task is set up to first copy the "oldcluster" pgBackRest
repository to a new PVC that will be used by "newcluster". We can do
this as all writes to the pgBackRest repository are atomic. We leverage
a Kubernetes Job to perform this task, which itself leverages rsync.

3. Once the pgBackRest repository is clone, we then perform a pgBackRest
restore operation using the newly created pgBackRest repository onto the
PVC where the data directory will reside for the "newcluster". This
brings over the essential files for running the new PostgreSQL cluster.

4. Next we launch a deployment of the new PostgreSQL cluster, which
becomes a warm standby and recovers.

5. Once the recovery is complete, our clone is ready to be used.

This has been verified using the "local" method of storing pgBackRest
data, but not the "s3" method. However, based on the present
implementation of the "s3" method, this should presumably work, but
would require further investigation that is presently out of scope.

Issue: [ch6609]