This commit adds support for replica 'start failed' detection, as implemented in the 'crunchy-postges-ha' container.  Specifically, by setting the PGHA_REPLICA_REINIT_ON_START_FAIL environment variable to true in the 'crunchy-postgres-ha' container for any PG deployment, the container will check for a 'start failed' state in a replica, and if detected will reinitialize the replica using the "reinitialize" REST endpoint deployed by Patroni on the replica.
Additionally, this commit adds additional support
for the 'start failed' detection process within the
PGO itself by providing supporting functionality
for the 'on_role_change' Patroni callback, i.e. taking
certain actions post-failover after detecting that
the role of a PG database has changed from 'replica'
to 'master', indicating promotion of a replica to
primary due to a failover (or switchover).  This
includes updating the pgBackRest dedicated
repository to point to the PGDATA directory of the
new/promoted primary, as well as taking a new
backup of the database.  The creation of a new
backup has the added benefit of ensuring any
replicas created or re-initialized post-failover can
successfully bootstrap (e.g. in the event that WAL
is missing from the pgBackRest repo, resulting in
the inability of a replica to otherwise recover).

While enabled by default, the replica 'start failed'
detection can be disabled by setting
DisableReplicaStartFailReinit to "true" in pgo.yaml.
And in addition to the above, also included in this
commit are are updated readiness probes for both
the database and crunchyadm containers.
Specifically, both now utilize the new readiness
scripts provided in the 'crunchy-postgres-ha'
container.